Only hash file sets with equal sizes
This adds a period where the program is just collecting files which slows it
down. But it only has to read a fraction of the data providing a huge speedup.

Hashed 37814 files, totalling 93388MB, and identified 11006 duplicates in 160034ms at 583MBps
Hashed 17476 files, totalling 20087MB, and identified 11006 duplicates in 43060ms at 466MBps

Signed-off-by: Tad <tad@spotco.us>